Update: root dir and onlyif attr added

+++ 0x0A-configuration_management/2-execute_a_command.pp
@@ -2,5 +2,6 @@
 
 exec{ 'killmenow':
   command => 'pkill killmenow',
-  path    => 'usr/bin',
-} 
+  path    => '/usr/bin',
+  onlyif  => 'pgrep killmenow',
+}
